+++ test/neatest.cpp
@@ -1,0 +1,8 @@
+/**
+ * @file      neatest.cpp
+ * @brief     The macro framework of the test suite.
+ * @copyright (c) Jörg Karl-Heinz Walter Brüggmann, 2021-2026
+ * @author    Jörg Karl-Heinz Walter Brüggmann <info@joerg-brueggmann.de>
+ */
+
+#include "neatest.h"