+++ test/test_widget.h
@@ -1,0 +1,15 @@
+/**
+ * @file      test_widget.h
+ * @brief     The test group of the component genc3wb::widget.
+ * @copyright (c) Jörg Karl-Heinz Walter Brüggmann, 2021-2026
+ * @author    Jörg Karl-Heinz Walter Brüggmann <info@joerg-brueggmann.de>
+ */
+
+#ifndef TEST_WIDGET_H
+#define TEST_WIDGET_H
+
+#include "neatest.h"
+
+TEST_DEF( widget );
+
+#endif // TEST_WIDGET_H